+++ workers/ohsome_quality_analyst/api/api.py
@@ -1,10 +1,8 @@
 import logging
 from typing import Optional
 
-import geojson
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from geojson import Feature, FeatureCollection, MultiPolygon, Polygon
 
 from ohsome_quality_analyst import __version__ as oqt_version
 from ohsome_quality_analyst import oqt
@@ -18,8 +16,7 @@ from ohsome_quality_analyst.api.request_models import (
     ReportRequestModel,
 )
 from ohsome_quality_analyst.geodatabase import client as db_client
-from ohsome_quality_analyst.utils.definitions import GEOM_SIZE_LIMIT, configure_logging
-from ohsome_quality_analyst.utils.helper import name_to_lower_camel
+from ohsome_quality_analyst.utils.definitions import configure_logging
 
 configure_logging()
 logging.info("Logging enabled")
@@ -35,47 +32,14 @@ app.add_middleware(
 )
 
 
-def empty_api_response() -> dict:
+def empty_api_response(request_url: str) -> dict:
     return {
-        "attribution": {
-            "url": "https://ohsome.org/copyrights",
-            "text": "© OpenStreetMap contributors",
-        },
         "apiVersion": oqt_version,
+        "attribution": {
+            "text": "© OpenStreetMap contributors",
+            "url": "https://ohsome.org/copyrights",
+        },
     }
-
-
-async def load_bpolys(bpolys: str) -> Feature:
-    """Load as GeoJSON object, validate and check size of bpolys"""
-
-    async def check_geom_size(geom):
-        if await db_client.get_area_of_bpolys(geom) > GEOM_SIZE_LIMIT:
-            raise ValueError(
-                "Input GeoJSON Object is too big. "
-                + "The area should be less than {0} sqkm.".format(GEOM_SIZE_LIMIT)
-            )
-
-    bpolys = geojson.loads(bpolys)
-
-    if bpolys.is_valid is False:
-        raise ValueError("Input geometry is not valid")
-    elif isinstance(bpolys, FeatureCollection):
-        if len(bpolys.features) == 1:
-            feature = bpolys.features[0]
-            await check_geom_size(feature.geometry)
-            return feature
-        else:
-            raise ValueError("Only one Feature is supported in a FeatureCollection")
-    elif isinstance(bpolys, Feature):
-        await check_geom_size(bpolys.geometry)
-        return bpolys
-    elif isinstance(bpolys, (Polygon, MultiPolygon)):
-        await check_geom_size(bpolys)
-        return Feature(geometry=bpolys)
-    else:
-        raise ValueError(
-            "Input GeoJSON Objects have to be of type Feature, Polygon or MultiPolygon"
-        )
 
 
 @app.get("/indicator/{name}")
@@ -143,22 +107,18 @@ async def _fetch_indicator(
     dataset: Optional[str] = None,
     feature_id: Optional[str] = None,
     fid_field: Optional[str] = None,
-):
-    if bpolys is not None:
-        feature = await load_bpolys(bpolys)
-    else:
-        feature = None
-
-    indicator = await oqt.create_indicator(
-        name, layer_name, feature, dataset, feature_id, fid_field
+) -> dict:
+    geojson_object = await oqt.create_indicator_as_geojson(
+        name,
+        layer_name,
+        bpolys,
+        dataset,
+        feature_id,
+        fid_field,
+        size_restriction=True,
     )
-
     response = empty_api_response()
-    response["metadata"] = vars(indicator.metadata)
-    response["metadata"].pop("result_description", None)
-    response["metadata"].pop("label_description", None)
-    response["layer"] = vars(indicator.layer)
-    response["result"] = vars(indicator.result)
+    response.update(geojson_object)
     return response
 
 
@@ -183,13 +143,7 @@ async def get_report(
         dataset = dataset.value
     if fidField is not None:
         fidField = fidField.value
-    return await _fetch_report(
-        name.value,
-        bpolys,
-        dataset,
-        featureId,
-        fidField,
-    )
+    return await _fetch_report(name.value, bpolys, dataset, featureId, fidField)
 
 
 @app.post("/report/{name}")
@@ -227,38 +181,16 @@ async def _fetch_report(
     feature_id: Optional[str] = None,
     fid_field: Optional[str] = None,
 ):
-    if bpolys is not None:
-        feature = await load_bpolys(bpolys)
-    else:
-        feature = None
-
-    report = await oqt.create_report(
+    geojson_object = await oqt.create_report_as_geojson(
         name,
-        feature=feature,
+        bpolys=bpolys,
         dataset=dataset,
         feature_id=feature_id,
         fid_field=fid_field,
+        size_restriction=True,
     )
-
     response = empty_api_response()
-    response["metadata"] = vars(report.metadata)
-    response["metadata"].pop("label_description", None)
-    response["result"] = vars(report.result)
-    response["result"]["label"] = report.result.label
-    response["indicators"] = {}
-    for indicator in report.indicators:
-        metadata = vars(indicator.metadata)
-        metadata.pop("result_description", None)
-        metadata.pop("label_description", None)
-        layer = vars(indicator.layer)
-        result = vars(indicator.result)
-        indicator_name = name_to_lower_camel(metadata["name"])
-        layer_name = name_to_lower_camel(layer["name"])
-        response["indicators"][indicator_name + layer_name] = {
-            "metadata": metadata,
-            "layer": layer,
-            "result": result,
-        }
+    response.update(geojson_object)
     return response