+++ workers/tests/integrationtests/test_api_indicator_geojson_io.py
@@ -1,0 +1,103 @@
+"""
+Testing FastAPI Applications:
+https://fastapi.tiangolo.com/tutorial/testing/
+"""
+# API request tests are seperated for indicator and report
+# because of a bug when using two schemata.
+
+import os
+import unittest
+from typing import Tuple
+
+from fastapi.testclient import TestClient
+from schema import Schema
+
+from ohsome_quality_analyst.api import app
+
+from .api_response_schema import (
+    get_feature_schema,
+    get_featurecollection_schema,
+    get_response_schema,
+)
+from .utils import oqt_vcr
+
+
+def get_fixture(name):
+    path = os.path.join(os.path.dirname(os.path.abspath(__file__)), "fixtures", name)
+    with open(path, "r") as f:
+        return f.read()
+
+
+class TestApiIndicatorIo(unittest.TestCase):
+    def setUp(self):
+        self.client = TestClient(app)
+
+        self.indicator_name = "GhsPopComparisonBuildings"
+        self.layer_name = "building_count"
+
+        self.response_schema = get_response_schema()
+        self.feature_schema = get_feature_schema()
+        self.featurecollection_schema = get_featurecollection_schema()
+
+    def run_tests(self, response, schemata: Tuple[Schema]) -> None:
+        self.assertEqual(response.status_code, 200)
+        for schema in schemata:
+            self.validate(response.json(), schema)
+
+    def validate(self, geojson: dict, schema: Schema) -> None:
+        schema.validate(geojson)  # Print information if validation fails
+        self.assertTrue(schema.is_valid(geojson))
+
+    def get_response(self, bpoly):
+        """Return HTTP GET response"""
+        url = "/indicator/{0}?layerName={1}&bpolys={2}".format(
+            self.indicator_name,
+            self.layer_name,
+            bpoly,
+        )
+        return self.client.get(url)
+
+    def post_response(self, bpoly):
+        """Return HTTP POST response"""
+        data = {"bpolys": bpoly, "layerName": self.layer_name}
+        url = f"/indicator/{self.indicator_name}"
+        return self.client.post(url, json=data)
+
+    @oqt_vcr.use_cassette()
+    def test_indicator_bpolys_geometry(self):
+        geometry = get_fixture("heidelberg-altstadt-geometry.geojson")
+        for response in (self.get_response(geometry), self.post_response(geometry)):
+            self.run_tests(response, (self.response_schema, self.feature_schema))
+
+    @oqt_vcr.use_cassette()
+    def test_indicator_bpolys_feature(self):
+        feature = get_fixture("heidelberg-altstadt-feature.geojson")
+        for response in (self.get_response(feature), self.post_response(feature)):
+            self.run_tests(response, (self.response_schema, self.feature_schema))
+
+    @oqt_vcr.use_cassette()
+    def test_indicator_bpolys_featurecollection(self):
+        featurecollection = get_fixture(
+            "heidelberg-bahnstadt-bergheim-featurecollection.geojson",
+        )
+        for response in (
+            self.get_response(featurecollection),
+            self.post_response(featurecollection),
+        ):
+            self.run_tests(
+                response, (self.response_schema, self.featurecollection_schema)
+            )
+            for feature in response.json()["features"]:
+                self.validate(feature, self.feature_schema)
+
+    @oqt_vcr.use_cassette()
+    def test_indicator_bpolys_size_limit(self):
+        feature = get_fixture("europe.geojson")
+        with self.assertRaises(ValueError):
+            self.get_response(feature)
+        with self.assertRaises(ValueError):
+            self.post_response(feature)
+
+
+if __name__ == "__main__":
+    unittest.main()