+++ workers/tests/integrationtests/utils.py
@@ -2,7 +2,6 @@ import inspect
 import os
 
 import vcr
-from schema import Optional, Or, Schema
 
 TEST_DIR = os.path.dirname(os.path.abspath(__file__))
 FIXTURE_DIR = os.path.join(TEST_DIR, "fixtures", "vcr_cassettes")
@@ -21,78 +20,4 @@ oqt_vcr = vcr.VCR(
     record_mode=os.getenv("VCR_RECORD_MODE", default="new_episodes"),
     match_on=["method", "scheme", "host", "port", "path", "query", "body"],
     func_path_generator=filename_generator,
-)
-
-
-api_schema_indicator = Schema(
-    {
-        "attribution": {
-            "url": str,
-            "text": str,
-        },
-        "apiVersion": str,
-        "metadata": {
-            "name": str,
-            "description": str,
-        },
-        "layer": {
-            "name": str,
-            "description": str,
-            "endpoint": str,
-            "filter": str,
-            Optional("ratio_filter"): Or(str, None),
-        },
-        "result": {
-            "timestamp_oqt": str,
-            "timestamp_osm": Or(str, None),
-            "value": Or(float, None),
-            "label": str,
-            "description": str,
-            "svg": str,
-            Optional("data"): Or(dict, None),
-        },
-    }
-)
-
-api_schema_report = Schema(
-    {
-        "attribution": {
-            "url": str,
-            "text": str,
-        },
-        "apiVersion": str,
-        "metadata": {
-            "name": str,
-            "description": str,
-        },
-        "result": {
-            "value": float,
-            "label": str,
-            "description": str,
-        },
-        "indicators": {
-            str: {
-                "metadata": {
-                    "name": str,
-                    "description": str,
-                },
-                "layer": {
-                    "name": str,
-                    "description": str,
-                    "endpoint": str,
-                    "filter": str,
-                    Optional("ratio_filter"): Or(str, None),
-                },
-                "result": {
-                    "timestamp_oqt": str,
-                    "timestamp_osm": Or(str, None),
-                    "value": Or(float, None),
-                    "label": str,
-                    "description": str,
-                    "svg": str,
-                    Optional("data"): Or(dict, None),
-                },
-            }
-        },
-    }
 )